# api/src/__generated__/openapi.yaml
openapi: 3.0.0
paths:
...
  /owners/{ownername}/projects/{project}/frames:
    post:
      description: Frameを新規作成します
      operationId: FramesController_createFrame
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            example: user1
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            example: project1
            type: string
      requestBody:
        required: true
        description: Frameの情報
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                description:
                  type: string
                  minLength: 1
                icon:
                  type: string
                  minLength: 1
                attribute:
                  type: object
                  properties:
                    name:
                      type: string
                      minLength: 1
                    type:
                      type: string
                      enum:
                        - text
                        - number
                        - boolean
                        - date
                        - richText
                        - map
                        - list
                  required:
                    - name
                    - type
              required:
                - name
                - description
                - icon
                - attribute
      responses:
        '200':
          description: Frameの作成に成功しました
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                  data:
                    type: object
                    properties:
                      frameId:
                        type: string
                        format: uuid
                    required:
                      - frameId
                required:
                  - statusCode
                  - message
                  - data
      security: &ref_6
        - supabase-auth: []
      summary: FramesController_createFrame / Frame新規作成
      tags: &ref_7
        - frames
  /owners/{ownername}/projects/{project}/frames/{frameId}:
    delete:
      description: Frameを削除します
      operationId: FramesController_deleteFrame
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            example: user1
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            example: project1
            type: string
        - name: frameId
          required: true
          in: path
          description: Frame ID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        '200':
          description: Frameの削除に成功しました
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: *ref_6
      summary: FramesController_deleteFrame / Frame削除
      tags: *ref_7
  /owners/{ownername}/projects/{project}/frames/{frameId}/description:
    put:
      description: Frameの説明を更新します
      operationId: FramesController_updateFrameDescription
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            example: user1
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            example: project1
            type: string
        - name: frameId
          required: true
          in: path
          description: Frame ID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      requestBody:
        required: true
        description: Frameの説明
        content:
          application/json:
            schema:
              type: object
              properties:
                description:
                  type: string
                  minLength: 1
              required:
                - description
      responses:
        '200':
          description: Frameの説明の更新に成功しました
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: *ref_6
      summary: FramesController_updateFrameDescription / Frame説明更新
      tags: *ref_7
  /owners/{ownername}/projects/{project}/frames/{frameId}/name:
    put:
      description: Frameの名前を更新します
      operationId: FramesController_updateFrameName
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            example: user1
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            example: project1
            type: string
        - name: frameId
          required: true
          in: path
          description: Frame ID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      requestBody:
        required: true
        description: Frameの名前
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
              required:
                - name
      responses:
        '200':
          description: Frameの名前の更新に成功しました
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: *ref_6
      summary: FramesController_updateFrameName / Frame名前更新
      tags: *ref_7
  /owners/{ownername}/projects/{project}/frames/{frameId}/attributes:
    put:
      description: Frameの属性を更新します
      operationId: FramesController_updateFrameAttributes
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            example: user1
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            example: project1
            type: string
        - name: frameId
          required: true
          in: path
          description: Frame ID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      requestBody:
        required: true
        description: Frameの属性
        content:
          application/json:
            schema:
              type: object
              properties:
                attributes:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                        minLength: 1
                      type:
                        type: string
                        enum:
                          - text
                          - number
                          - boolean
                          - date
                          - richText
                          - map
                          - list
                      id:
                        type: string
                        format: uuid
                      description:
                        type: string
                    required:
                      - name
                      - type
                      - id
              required:
                - attributes
      responses:
        '200':
          description: Frameの属性の更新に成功しました
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: *ref_6
      summary: FramesController_updateFrameAttributes / Frame属性更新
      tags: *ref_7
  /owners/{ownername}/projects/{project}/frames/{frameId}/templates:
    post:
      description: Frameのテンプレートを作成します
      operationId: FramesController_createFrameTemplate
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            example: user1
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            example: project1
            type: string
        - name: frameId
          required: true
          in: path
          description: Frame ID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      requestBody:
        required: true
        description: Frameのテンプレート
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
              required:
                - name
      responses:
        '200':
          description: Frameのテンプレートの作成に成功しました
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: *ref_6
      summary: FramesController_createFrameTemplate / Frameテンプレート作成
      tags: *ref_7
  /owners/{ownername}/projects/{project}/frames/{frameId}/templates/{templateId}:
    put:
      description: Frameのテンプレートを更新します
      operationId: FramesController_updateFrameTemplate
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            example: user1
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            example: project1
            type: string
        - name: frameId
          required: true
          in: path
          description: Frame ID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
        - name: templateId
          required: true
          in: path
          description: Template ID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      requestBody:
        required: true
        description: Frameのテンプレート
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  minLength: 1
                content:
                  oneOf:
                    - oneOf:
                        - type: string
                        - type: number
                        - type: boolean
                        - {}
                    - type: array
                      items: {}
                    - type: object
                      additionalProperties: {}
              required:
                - name
                - content
            examples:
              default:
                value:
                  name: template
                  content:
                    type: doc
                    content: []
      responses:
        '200':
          description: Frameのテンプレートの更新に成功しました
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: *ref_6
      summary: FramesController_updateFrameTemplate / Frameテンプレート更新
      tags: *ref_7
    delete:
      description: Frameのテンプレートを削除します
      operationId: FramesController_deleteFrameTemplate
      parameters:
        - name: ownername
          required: true
          in: path
          description: Owner名
          schema:
            example: user1
            type: string
        - name: project
          required: true
          in: path
          description: Project名
          schema:
            example: project1
            type: string
        - name: frameId
          required: true
          in: path
          description: Frame ID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
        - name: templateId
          required: true
          in: path
          description: Template ID
          schema:
            example: 123e4567-e89b-12d3-a456-426614174000
            type: string
      responses:
        '200':
          description: Frameのテンプレートの削除に成功しました
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    description: ステータスコード
                    type: number
                  message:
                    description: メッセージ
                    type: string
                required:
                  - statusCode
                  - message
      security: *ref_6
      summary: FramesController_deleteFrameTemplate / Frameテンプレート削除
      tags: *ref_7
...
info:
  title: andTopic API
  description: andTopic API description
  version: '1.0'
  contact: {}
...
servers: []
...